Removed non-needed callbacks manipulation on conn connection
We didn't need to create a different callback for each socket connected to a connection,
we already had a Socket.io Room to manage this.

So, this "one-socket-one-callback feature" was removed, but we still saves the socket id on a connection,
which enables the connection to know how many clients are connected to it.